updated README.md, seed_tables.sql, and setup_tables.sql

+++ setup_tables.sql
@@ -9,14 +9,14 @@ $$ LANGUAGE plpgsql;
 
 CREATE TABLE IF NOT EXISTS results (
     id serial PRIMARY KEY,
-	user_id int UNIQUE,
-	color VARCHAR ( 100 ),
+        user_id BIGINT UNIQUE,
+        color VARCHAR ( 100 ),
     created_at TIMESTAMP NOT NULL DEFAULT NOW(),
     updated_at TIMESTAMP NOT NULL DEFAULT NOW()
 );
 
 
 CREATE TRIGGER set_timestamp
-BEFORE UPDATE ON results 
+BEFORE UPDATE ON results
 FOR EACH ROW
 EXECUTE PROCEDURE trigger_set_timestamp();
